requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
celery==5.3.0
redis==4.6.0
easyocr==1.6.2
pypdfium2==4.0.0
pillow==10.0.0
python-multipart==0.0.6
pandas>=1.5
rapidfuzz>=3.0